Add files via upload
This project involved designing and implementing a noise-canceling headphone system using adaptive filters as part of a Digital Signal Processing course. We successfully modeled the secondary path effects to account for the non-idealities of real-world components. However, the main adaptive filter code is still under troubleshooting, with plans to resolve the issue and achieve a fully functional system.